requirements.txt
decorator==4.4.2
networkx==2.5
PyYAML==5.3.1
